refactor(proxy): 拆分代理执行模块并缩短路由关键路径
