+++ CONTEXT.md
@@ -1,0 +1,8 @@
+# AutoRouter 代理上下文
+
+本上下文统一描述 AutoRouter 代理请求的路由与执行概念，避免把网关决策、上游响应和客户端体验混为一谈。
+
+## 代理路由
+
+**路由决策延迟**：代理请求进入网关后，到开始向选定上游发起请求之前，认证、准入、候选构建、能力匹配、熔断检查、排队与选路所消耗的时间。
+_避免_：总响应延迟、上游响应延迟、TTFT
